Renamed non-api api files from route to something else for temp fix.
