+++ app/api/data/school-data/get-school-data.ts
